requirements.txt
fastapi
uvicorn
gunicorn
torch
torchvision
transformers
pillow
numpy
requests